requirements.txt
django
google-generativeai
edge-tts
beautifulsoup4
requests
python-dotenv
moviepy
Pillow
numpy
